requirements.txt
streamlit>=1.35.0
ollama>=0.2.1
feedparser>=6.0.11
beautifulsoup4>=4.12.3
requests>=2.31.0
pandas>=2.2.0
fpdf2>=2.7.9
pyyaml>=6.0.1
lxml>=5.2.1
python-dateutil>=2.9.0
# Novas dependências
APScheduler>=3.10.4
python-docx>=1.1.2
plotly>=5.22.0
